+++ Changelog.md
@@ -21,6 +21,9 @@
 - Clojure: Improves error and warning messages. ([#813](https://github.com/fossas/fossa-cli/pull/813))
 - Nim: Improves error and warning messages. ([#813](https://github.com/fossas/fossa-cli/pull/813))
 - Rust: Improves error and warning messages. ([#813](https://github.com/fossas/fossa-cli/pull/813))
+- UX: Improves errors for dynamic deps, and binary deps analysis. ([#819](https://github.com/fossas/fossa-cli/pull/819))
+- UX: Improves analysis scan summary rendering. ([#819](https://github.com/fossas/fossa-cli/pull/819))
+
 
 ## v3.1.0